save the things

+++ lib/vagrant-zones/driver.rb
@@ -87,34 +87,49 @@ module VagrantPlugins
 				name = @machine.name
 				machine.config.vm.networks.each do |_type, opts|
 					responses=[]
+					nic_number	= opts[:nic_number].to_s
+					if !opts[:type].nil?
+						nictype  = opts[:nictype]
+					end
+					case nictype
+					when /external/
+					  nic_type = "e"
+					when /internal/
+					  nic_type = "i"
+					when /carp/
+					  nic_type = "c"
+					when /management/
+					  nic_type = "m"
+					when /host/
+					  nic_type = "h"
+					else
+					  nic_type = "e"
+					end
 					if _type.to_s == "public_network"
 						if opts[:dhcp] == true
-							PTY.spawn("pfexec zlogin -C #{name}") do |zlogin_read,zlogin_write,pid|
-								zlogin_read.expect(/\n/) { |msg| zlogin_write.printf("hostname -I\n") }
-								Timeout.timeout(30) do
-									loop do
-										zlogin_read.expect(/\r\n/) { |line|  responses.push line}
-										if responses[-1].to_s.match(/^(?:[0-9]{1,3}\.){3}[0-9]{1,3}$/)
-											ip = responses[-1][0].rstrip.gsub(/\e\[\?2004l/, "").lstrip
-											return nil if ip.length == 0
-											return ip.gsub /\t/, ''
-											break
-										elsif responses[-1].to_s.match(/Error Code: \b(?![0]\b)\d{1,4}\b/)
-												raise "==> #{name} ==> Command ==> #{cmd} \nFailed with ==> #{responses[-1]}"
+							if opts[:managed]
+								PTY.spawn("pfexec zlogin -C #{name}") do |zlogin_read,zlogin_write,pid|
+									zlogin_read.expect(/\n/) { |msg| zlogin_write.printf("ip -4 addr show dev vnic#{nic_type}#{machine.config.vm_type}_#{machine.partition_id}_#{nic_number} | head -n -1 | tail -1  | awk '{ print $2 }'\n") }
+									Timeout.timeout(30) do
+										loop do
+											zlogin_read.expect(/\r\n/) { |line|  responses.push line}
+											puts responses[-1]
+											if responses[-1].to_s.match(/^(?:[0-9]{1,3}\.){3}[0-9]{1,3}$/)
+												ip = responses[-1][0].rstrip.gsub(/\e\[\?2004l/, "").lstrip
+												return nil if ip.length == 0
+												return ip.gsub /\t/, ''
+												break
+											elsif responses[-1].to_s.match(/Error Code: \b(?![0]\b)\d{1,4}\b/)
+													raise "==> #{name} ==> Command ==> #{cmd} \nFailed with ==> #{responses[-1]}"
+											end
 										end
 									end
+									Process.kill("HUP",pid)
 								end
-								Process.kill("HUP",pid)
 							end
 							puts "==> #{machine.name} ==> DHCP is not yet Configured for use"
 						elsif opts[:dhcp] == false
 							if opts[:managed]
-								puts
-								puts
-								puts 
-								puts opts[:ip].to_s
-								puts
-								puts
 								ip = opts[:ip].to_s
 								return nil if ip.length == 0
 								return ip.gsub /\t/, ''
@@ -279,6 +294,7 @@ end								}
       dhcp-identifier: mac
       dhcp4: yes
       dhcp6: no
+	  set-name: vnic#{nic_type}#{config.vm_type}_#{config.partition_id}_#{nic_number}
       nameservers:
         addresses: [#{nameserver1} , #{nameserver2}]	}
 														if run == 0
@@ -299,6 +315,7 @@ end								}
       dhcp-identifier: mac
       dhcp4: no
       dhcp6: no
+	  set-name: vnic#{nic_type}#{config.vm_type}_#{config.partition_id}_#{nic_number}
       addresses: [#{ip}/#{netmask}]
       gateway4: #{defrouter}
       nameservers:
